Camios en auth

+++ src/main/java/com/example/demo/service/ProductoService.java
@@ -16,5 +16,5 @@ public interface ProductoService {
 	public List<ProductoDeatllesDto> obtenerDetalles(Integer id);
 	public List<Producto> filtrar(Integer marca, Integer categoria, String color, Integer talla, Double precio_min, Double precio_max);
 	public List<Producto> filtrar(List<List<String>> ids);
-	public int obtenerDetalle(Integer id, String color, Integer talla, Integer cantidad);
+	public long obtenerDetalle(Integer id, String color, Integer talla, Integer cantidad);
 }
